Correct time after file has ended and overlapping play until no file in buffer

+++ src/sample_getter/audio_file.h
@@ -22,7 +22,7 @@ class AudioFile : public SampleGetter {
 
     private:
         float *wav_buffer;
-        unsigned long wav_buffer_samples;
+        int wav_buffer_samples;
         SDL_AudioSpec wav_spec;
 };
